requirements.txt
streamlit>=1.29.0
supabase>=2.0.0
openai>=1.0.0
pandas>=2.0.0
plotly>=5.18.0
python-dotenv>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.8.0
